Update Policy.java
back to int from byte for policyNum input for argPolicy

+++ Policy.java
@@ -23,7 +23,7 @@ public class Policy
       holderWeight = 0;
    }   
    
-   public void argPolicy(byte polNum,String provName,String holdFirst,String holdLast,byte holdAge,String smokeStatus,double holdHeight,double holdWeight)
+   public void argPolicy(int polNum,String provName,String holdFirst,String holdLast,byte holdAge,String smokeStatus,double holdHeight,double holdWeight)
    { //Constructor with arguments
       policyNum = polNum;
       providerName = provName;
